+++ Source/Module/AkronModuleSettings.cs
@@ -672,6 +672,7 @@ public partial class AkronModuleSettings : EverestModuleSettings {
     public bool ScreenshotScannerExportStartPositions { get; set; } = true;
     public bool ScreenshotScannerExportAutoKillAreas { get; set; } = true;
     public bool ScreenshotScannerExportAutoDeafenAreas { get; set; } = true;
+    public bool ScreenshotScannerDownscaleMapCapture { get; set; }
     public bool ScreenshotScannerFreezeTime { get; set; } = true;
     public bool ScreenshotScannerRemoveBackground { get; set; }
     public bool ScreenshotScannerRemoveForeground { get; set; }